enforce `cv` section to have `cv` title
The reason this hardcoding is acceptable is that `cv` section uses its own `cv` layout. In other words, `cv` is already indepedent of other sections, so we can simply check if a site page (ie sections such as publication) is using `cv` layout and if this is the case, enforce custom naming such as `cv` here.

All of these happened because I wanted to use `cv` title in navbar and `Curriculum Vitae` in the title of `cv.md`