clarify current student/job status
I am also confused in that regard, but to clarify, in Europe, PhDs are employees not students, hence the official legal title is "Doctoral Researcher" but I feel it says I already have a doctoral degree. As a result the closest terminology that could represent my highest degree is MSc yet I am doing research was "Research Associate" yet havent used it yet.